basic refactored code
