+++ src/manifest/graphMapper.ts
@@ -1,0 +1,86 @@
+import { Analysis, DBTGraphType, GraphMetaMap, Model, NodeGraphMap, NodeMetaMap, Seed, Source, SourceMetaMap, Test, Node } from "../domain";
+import { notEmpty } from "../utils";
+
+export class GraphMapper {
+    static createGraphMetaMap(
+        parentMap: DBTGraphType,
+        childrenMap: DBTGraphType,
+        modelMetaMap: NodeMetaMap,
+        sourceMetaMap: SourceMetaMap
+    ): GraphMetaMap {
+        const unique = (nodes: any[]) => Array.from(new Set(nodes));
+
+        const parents: NodeGraphMap = Object.entries(parentMap).reduce(
+            (map, [nodeName, nodes]) => {
+                const currentNodes = unique(nodes)
+                    .map(GraphMapper.mapToNode(sourceMetaMap, modelMetaMap))
+                    .filter(notEmpty);
+                map.set(nodeName, { nodes: currentNodes });
+                return map;
+            },
+            new Map()
+        );
+
+        const children: NodeGraphMap = Object.entries(childrenMap).reduce(
+            (map, [nodeName, nodes]) => {
+                const currentNodes = unique(nodes)
+                    .map(GraphMapper.mapToNode(sourceMetaMap, modelMetaMap))
+                    .filter(notEmpty);
+                map.set(nodeName, { nodes: currentNodes });
+                return map;
+            },
+            new Map()
+        );
+
+        return {
+            parents,
+            children,
+        };
+    }
+
+    static mapToNode(
+        sourceMetaMap: SourceMetaMap,
+        nodeMetaMap: NodeMetaMap
+    ): (parentNodeName: string) => Node | undefined {
+        return (parentNodeName) => {
+            const nodeSegment = parentNodeName.split(".");
+            const nodeType = nodeSegment[0];
+            switch (nodeType) {
+                case "source": {
+                    const sourceName = nodeSegment[2];
+                    const tableName = nodeSegment[3];
+                    const url = sourceMetaMap.get(sourceName)?.path!;
+                    return new Source(
+                        `${tableName} (${sourceName})`,
+                        parentNodeName,
+                        url
+                    );
+                }
+                case "model": {
+                    const modelName = nodeSegment[2];
+                    const url = nodeMetaMap.get(modelName)?.path!;
+                    return new Model(modelName, parentNodeName, url);
+                }
+                case "seed": {
+                    const modelName = nodeSegment[2];
+                    const url = nodeMetaMap.get(modelName)?.path!;
+                    return new Seed(modelName, parentNodeName, url);
+                }
+                case "test": {
+                    const modelName = nodeSegment[2];
+                    const url = nodeMetaMap.get(modelName)?.path!;
+                    return new Test(modelName, parentNodeName, url);
+                }
+                case "analysis": {
+                    const modelName = nodeSegment[2];
+                    const url = nodeMetaMap.get(modelName)?.path!;
+                    return new Analysis(modelName, parentNodeName, url);
+                }
+                default:
+                    console.log(`Node Type '${nodeType}' not implemented!`);
+                    return undefined;
+            }
+        };
+    }
+
+}
